[repo initialization] fix: disables pre-commits after repo creation
if a user has a global templatedir directive in
their git config it can result in an executable
pre-commit being installed when the repo is
initialized.

this is never what we want. PC will now make that
NOT executable and disable the git warning about
it in that repo.

if, for some strange reason, you _do_ want a
pre-commit hook on your private comments commits
then you can set the "PRIVATE_COMMENTS_ALLOW_PRE_COMMIT"
environment variable to "true"

this commit also addresses issue #14 which was a side-effect
of this bad pre-commit state

TICKET: https://github.com/masukomi/private_comments/issues/15
FIXES #15
FIXES #14

TAGS: hooks