feat: adds build for Linux
This change adds some affordances for building private_comments
under Linux.  Since I could not guarantee that every builder would
have all the necessary tools for building, I wrapped the build
process into a docker environment.

This allows an almost trivial build command line:

```
./linux/make.release.sh
```

You need to have docker installed for this to be possible, but it might
not be a big concession to make for the ease of building. This still
requires root access to run Docker (sigh), but doesn't require you to
install a whole new runtime if you are only interested in running the
binaries.

I used shell scripts for this purpose to follow the style established in
the repository, though I'd have prefered to use some `make` tool.

Issues: #22